+++ firmware/controllers/trigger/decoders/trigger_structure.cpp
@@ -159,7 +159,10 @@ angle_t TriggerWaveform::getAngle(int index) const {
 	int crankCycle = index / privateTriggerDefinitionSize;
 	int remainder = index % privateTriggerDefinitionSize;
 
-	return getCycleDuration() * crankCycle + getSwitchAngle(remainder);
+	auto cycleStartAngle = getCycleDuration() * crankCycle;
+	auto positionWithinCycle = getSwitchAngle(remainder);
+
+	return cycleStartAngle + positionWithinCycle;
 }
 
 void TriggerWaveform::addEventClamped(angle_t angle, trigger_wheel_e const channelIndex, trigger_value_e const stateParam, float filterLeft, float filterRight) {
@@ -210,6 +213,10 @@ void TriggerWaveform::calculateExpectedEventCounts(bool useOnlyRisingEdgeForTrig
 
 void TriggerWaveform::addEvent720(angle_t angle, trigger_wheel_e const channelIndex, trigger_value_e const state) {
 	addEvent(angle / 720, channelIndex, state);
+}
+
+void TriggerWaveform::addEvent360(angle_t angle, trigger_wheel_e const channelIndex, trigger_value_e const state) {
+	addEvent(CRANK_MODE_MULTIPLIER * angle / 720, channelIndex, state);
 }
 
 void TriggerWaveform::addEventAngle(angle_t angle, trigger_wheel_e const channelIndex, trigger_value_e const state) {
@@ -362,7 +369,7 @@ void TriggerWaveform::setTriggerSynchronizationGap3(int gapIndex, float syncRati
 /**
  * this method is only used on initialization
  */
-uint16_t TriggerWaveform::findAngleIndex(TriggerFormDetails *details, float target) const {
+uint16_t TriggerWaveform::findAngleIndex(TriggerFormDetails *details, angle_t targetAngle) const {
 	size_t engineCycleEventCount = getLength();
 
 	efiAssert(CUSTOM_ERR_ASSERT, engineCycleEventCount != 0 && engineCycleEventCount <= 0xFFFF, "engineCycleEventCount", 0);
@@ -378,9 +385,9 @@ uint16_t TriggerWaveform::findAngleIndex(TriggerFormDetails *details, float targ
         int middle = (left + right) / 2;
 		angle_t eventAngle = details->eventAngles[middle];
 
-        if (eventAngle < target) {
+        if (eventAngle < targetAngle) {
             left = middle + 1;
-        } else if (eventAngle > target) {
+        } else if (eventAngle > targetAngle) {
             right = middle - 1;
         } else {
             // Values are equal
@@ -414,7 +421,10 @@ void findTriggerPosition(TriggerWaveform *triggerShape,
 
 	int triggerEventIndex = details->triggerIndexByAngle[(int)angle];
 	angle_t triggerEventAngle = details->eventAngles[triggerEventIndex];
-	if (angle < triggerEventAngle) {
+	angle_t offsetFromTriggerEvent = angle - triggerEventAngle;
+
+	// Guarantee that we aren't going to try and schedule an event prior to the tooth
+	if (offsetFromTriggerEvent < 0) {
 		warning(CUSTOM_OBD_ANGLE_CONSTRAINT_VIOLATION, "angle constraint violation in findTriggerPosition(): %.2f/%.2f", angle, triggerEventAngle);
 		return;
 	}
@@ -424,7 +434,7 @@ void findTriggerPosition(TriggerWaveform *triggerShape,
 		chibios_rt::CriticalSectionLocker csl;
 
 		position->triggerEventIndex = triggerEventIndex;
-		position->angleOffsetFromTriggerEvent = angle - triggerEventAngle;
+		position->angleOffsetFromTriggerEvent = offsetFromTriggerEvent;
 	}
 }
 
@@ -541,12 +551,16 @@ void TriggerWaveform::initializeTriggerWaveform(operation_mode_e ambiguousOperat
 		configureFordAspireTriggerWaveform(this);
 		break;
 
-	case TT_VVT_NISSAN_VQ:
+	case TT_VVT_NISSAN_VQ35:
 		initializeNissanVQvvt(this);
 		break;
 
-	case TT_TT_NISSAN_VQ:
-		initializeNissanVQcrank(this);
+	case TT_NISSAN_VQ30:
+		initializeNissanVQ30cam(this);
+		break;
+
+	case TT_NISSAN_VQ35:
+		initializeNissanVQ35crank(this);
 		break;
 
 	case TT_KAWA_KX450F: