+++ firmware/controllers/trigger/decoders/trigger_structure.h
@@ -224,6 +224,11 @@ public:
 	void addEvent720(angle_t angle, trigger_wheel_e const channelIndex, trigger_value_e const state);
 
 	/**
+	 * this method helps us use real world 350 degrees shape for FOUR_STROKE_CAM_SENSOR and FOUR_STROKE_CRANK_SENSOR
+	 */
+	void addEvent360(angle_t angle, trigger_wheel_e const channelIndex, trigger_value_e const state);
+
+	/**
 	 * This version of 'addEvent...' family considers the angle duration of operationMode in this trigger
 	 * For example, (0..180] for FOUR_STROKE_SYMMETRICAL_CRANK_SENSOR
 	 */
@@ -271,10 +276,11 @@ public:
 			const trigger_config_s& triggerConfig
 			);
 
+	uint16_t findAngleIndex(TriggerFormDetails *details, angle_t angle) const;
+
 private:
 	trigger_shape_helper h;
 
-	uint16_t findAngleIndex(TriggerFormDetails *details, float angle) const;
 
 	/**
 	 * Working buffer for 'wave' instance