+++ firmware/gen_config.sh
@@ -17,10 +17,11 @@ bash gen_config_default.sh
 # maybe one day we will automate but not yet
 #
 for BOARD in "f429-discovery f429-discovery" "hellen/hellen128 hellen128" "hellen/hellen121vag hellen121vag" "hellen/hellen121nissan hellen121nissan" "hellen/hellen72 hellen72" "hellen/hellen64_miataNA6_94 hellenNA6" "microrusefi mre_f7" "microrusefi mre_f4" "frankenso frankenso_na6" "prometheus prometheus_469" "prometheus prometheus_405" "proteus proteus_f7" "proteus proteus_f4"; do
- BOARD_NAME="${BOARD% *}"
- BOARD_SHORT_NAME="${BOARD#* }"
- bash gen_config_board.sh $BOARD_NAME $BOARD_SHORT_NAME
- [ $? -eq 0 ] || { echo "ERROR generating board $BOARD_NAME $BOARD_SHORT_NAME"; exit 1; }
+ BOARD_NAME=$(echo "$BOARD" | cut -d " " -f 1)
+ BOARD_SHORT_NAME=$(echo "$BOARD" | cut -d " " -f 2)
+ INI=$(echo "$BOARD" | cut -d " " -f 3)
+ bash gen_config_board.sh $BOARD_NAME $BOARD_SHORT_NAME $INI
+ [ $? -eq 0 ] || { echo "ERROR generating board $BOARD_NAME $BOARD_SHORT_NAME $INI"; exit 1; }
 done
 
 cd config/boards/kinetis/config